Revert "Migrate to App Router (#1015)"
This reverts commit a6efecfc25e5cfd6d976788c0f0d5cf7b462e325.
