+++ app/src/pages/index.tsx
@@ -1,3 +1,2 @@
-'use client';
 import { Redirect } from '../lib/redirect';
 export default Redirect;
